move cmd in a dedicated package
